Allow adding Fields to GenericObject
Modified method "getItemtypes" and added two new static methods. The goal is to allow users to associate custom fields to generic objects. For context and relevance, see tickets 197, 23, 83 and 188 in github.com/pluginsGLPI/fields. In ticket 197, author "trasher" briefly explains why a general solution is not yet available. This proposed solution is limited to adding fields to objects created with the GenericObject plugin and is designed to fail quickly when assumptions fail.