Remove hack to link Fields to GenericObject
After thorough testing, the hack that allows the Fields plugin to add custom fields to GenericObjects is to much trouble.

Plugin Fields expects objects to be eternal, because it was designed to work with standard GLPI objects. The full life cycle of GenericObjects is not handled by Fields and is out of scope of the Fields plugins.

Moreover, the underlying requirement is to add a GUI for managing plugin GenericObjects' own solution for custom fields. This is the recommended approach and avoids coupling these two plugins.

Another alternative would be a new plugin that merges the functionnalities of both plugins in a single unified solution.